feat: pass `fetchOptions` to fetch client
This PR replaces the previous keepalive behavior with a way to override
all options passed to `fetch()`. The benefit here is that additional
headers that a user might require are available.

Since we didn't post a release on npm, we will not treat this as an ABI breaking version.

PR: https://github.com/Topsort/topsort.js/pull/170