fix: export esm as module
In a previous release, we switched from the package being commonjs
to being ESM. This also changed the default behavior in `tsup`, which
led to the currently exported `index.js` "incorrectly" being ESM.

In this commit, we adjust this by properly changing the export.

While at it, simplify the test suite to not use IIFE and finally
migrate the bundle stack to use bunup which is both faster and
generates faster and smaller output.